requirements.txt
Django==4.2.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
pymongo==4.6.1